Introduce the concept of catalog DB connections.
Formerly we used a fixed set of database connection
parameters per archive. With these changes, we can
create and manage connections separately from an
existing archive, which allows us, for example, to
create dedicated database connection strings for
streaming archive connections for dedicated connections.

This creates a new catalog table connections, where all
DB connection information is stored. Currently we have two
types of connections there: basebackup and streamer.

The parser now distinguishes between the old syntax, where
all keywords for a database connection has to be specified, or
the new DSN syntax, which allows to specify the database URI
directly. Currently we don't verify this against libpq, this
might be a further improvement.

With this commit, we also need to force FK checks to on in SQLite3.