Work on WALStreamerProcess API.
Improve implementation for receive() to stream WAL
records to an archive. Currently this is just a WIP, many
functionality is still missing. handleReceive() and friends
still hold many bugs, but i commit this for now to make any progress
on testing the infrastructure.

Introduce a new command START STREAMING FOR ARCHIVE,
which starts WAL streamer processes for the specified archive. This
is yet here for debugging purposes, but might persist later for
user interaction.

Reshape some of the logic in PGStream::createPhysicalReplicationSlot(),
since we need the archive_name somehow in the slot identifier. I'm not
really satisfied yet how this whole slot mess is handled, but given that
we want to have the slots to persist, we might have to go that route.