Introduce WAL segment handling and improve WALStreamerProcess.
ArchiveLogDirectory now provides functions to scan and
analyze the log/ directory of the archive, as well as helper
methods to return the uncompressed (*real*) size of
a WAL segment. We also support now to get a XLogRecPtr start
position from the physical WAL files via getXlogStartPosition().

The latter is required to startup a WALStreamerProcess iff
it had crashed somewhere, leaving an invalid xlog position
in the catalogs behind.

There's still a bug somewhere lurking around which prevents
the current implementation to enter a COPY stream of
XLOG messages, this is a topic for a followup commit to fix.