Proceed work on background launcher and worker infrastructure.
This commit introduces shared memory control areas
for the catalog launcher background process and worker processes
forked by this launcher. The LauncherSHM segment also controls
wether a launcher for the given catalog is already running, which
seems safer than just relying on the sqlite catalog. Only the
background launcher process is allowed being attached to this
control segment, which allocates just 4K in memory.

A WorkerSHM segment is introduced to control currently running background
processes. Currently this is a fixed size of 10M, but WorkerSHM::calculateSHMsize()
is already prepared to return the size depending on the current setting of
WorkerSHM::max_workers.
Every forked background worker is registered in this shared memory
segment. Currently this is just a list where each worker gets an index into
and stores its runtime information at this index. If a worker is terminated,
it is responsible to clean its slot afterwards.

The missing part here is if a background worker is killed or crashed, it will
leave behind an orphaned shared memory slot, which should be cleaned somehow.

The idea here is, that a further commit will introduce an advanced (SIGCHLD)
signal handler which reacts on the termination of a background worker. Since we
already track the worker PIDs in the worker shared memory area, we can
easily remove the orphaned entry.

There's another problem: If a launcher is killed or crashed, currently all
background workers will survive. I'm not comittet to a certain behavior here,
but it seems the future direction should be that in this case we just reattach
a new launcher instance. In case the launcher crashes or is SIGKILL'ed, we should
terminate all currently running background workers.