Introduce reaper infrastructure for dead background workers.
background_reaper is a base class to implement
reap() actions on PIDs from worker processes that died
or crashed somehow. Currently it acts on SIGKILL and
SIGABRT, but SIGSEGV and others might be interesting as well.

the background_worker_shm_reaper is designed to clean up
died workers from the worker shared memory area.

Also introduce a SHOW WORKER command to display current
active worker processes based on the current runtime information
found in the worker shared memory area.

This commit also fixes an issue with BackupCatalog::getStreams(), which
doesn't do the handling for return codes in sqlite3_step() right. It looks
like the whole handling needs to be reviewed again, but this will be
adressed by another commit then.