Finalize VERFIY ARCHIVE ... CONNECTION option.
When VERIFY ARCHIVE is executed with the CONNECTION option,
the command tests all basebackup and streaming connections
used by the archive. Currently we just call PGStream::testConnection(),
which employs PQPing(). This tests wether the specified database server
is reachable and has valid connection options, but doesn't actually
check the credentials.