Introduce STOP STREAMING FOR ARCHIVE command.
This makes it easier to control streaming workers from
the interactive pg_backup_ctl++ shell. STOP STREAMING FOR ARCHIVE
checks wether there is a running streaming background worker
for the specified archive and sends a SIGTERM to abort streaming
immediately. To accomplish this, commands.cxx needs access to
the shared memory handles, previously defined in jobhandles.hxx.

Since the latter also include commands.hxx, we need to separate
the shared memory handler definitions into a separate header, otherwise
we have the trouble of circular includes. shm.hxx is the new interface to
include when someone wants to access shared memory areas for
launcher and worker (via LauncherSHM and WorkerSHM). This also avoids
the problem of circular includes now.

Since we specify the archive name via STOP STREAMING FOR ARCHIVE, we
also need to track the archive identifier in the worker shared
memory area to identify the correct worker pid. This isn't straight
forward, since until now we don't have any tracking information for this.

To reach the desired behavior, PGBackupCtlCommand gets a new method
archive_name(), which returns the archive name of the command handle
it encapsulates (if not attached to a archive_name, an empty string
is returned). Having this allows the worker_command() routine
to query the backup catalog to obtain the archive ID we want to
store along with the other worker information in the worker shared
memory area.

This also requires a new interface method existsById() in the
BackupCatalog object, since we don't want to have a variable length
character string in shared memory and storing just the ID forces us
to implement an API method to query by ID.

While at it, change the implementation of
ShowWorkersCommandHandle::execute(). Previously we used to iterate
over the shared memory areas and print all information immediately
to STDOUT. Since i personally don't like any I/O during
critical sections (and it is a critical section since we hold a
mutex through the iteration), change to behavior to accumulate
the information into a vector and then printing the information
*after* releasing the mutex. This looks much safer.

I'm also planning to change the current mutex implementation from
a normal mutex into a scoped mutex, which gives better control
when a code block exits in an unforseen way.

This commit also adds some more commands to the command reference
documentation.