Move check for existing archive before instructing background launcher.
StartStreamingForArchiveCommand::execute() already checked for
the existence of the specified archive identifier. However, in DETACH
mode (which is the default), this never happened, since we already
sent the command through the message queue to the background
launcher. Though, the launched background process still failed, since
it would have reach the check then, but it seems nicer to
save the whole send command, fork() and error stuff.

To fix this, we move the check before communicating the START
STREAMING command to the background launcher.