Change the way we are going to apply PIN/UNPIN actions.
Previously the plan was to apply them in ascending order,
oldest backup to newest. But the old pg_backup_ctl script
does it in the opposite, newest to oldest. So change the
approach here to match the old behavior.

Also make the output for LIST BASEBACKUPS to display
the newest basebackups first, sorted by creation date
in descending order. This commit just adjusts the documentation,
code change will follow.