Finally implement PIN/UNPIN command actions.
This commit finalizes the PIN/UNPIN infrastructure
started with commit dbcaf24ab90e1b3392609ca695e7e5c67cc4b354. The
implementation supports the following concepts now:

PIN or UNPIN operations can be performed on a list of basebackups in batched
operations with the following actions:

PIN { +n | OLDEST | NEWEST | <basebackup ID> } IN ARCHIVE <identifier>
UNPIN { +n | OLDEST | NEWEST | <basebackup ID> | CURRENT } IN ARCHIVE <identifier>

OLDEST and NEWEST reference the oldest or newest basebackup in the
list of basebackups in the specified archive. The +n syntax is the number
of basebackups to be pinned or unpinned, starting with the newest one.
It is also possible to PIN or UNPIN a specific basebackup, referenced
by its backup ID which can be derived from LIST BASEBACKUPS.

In addition to PIN, UNPIN also supported the CURRENT action, which allows
to unpin all currently pinned basebackups in one go.

The PinCatalogCommand command instance is somehow special, compared to
the other BaseCatalogCommand implementations: it is used by two catalog
command tags at the same time and supports the PIN and UNPIN operation
within a single implementation of its execute() method. The information
on what to do is postponed into the PinRetention object instance, which
dispatches the operations to do based on the information encoded into the
PinDescr or UnpinDescr descriptor instances carried by the catalog descriptor.

The assignment operator and copy() method of CatalogDescr and BaseCatalogCommand
instance need to take care to copy the PinDescr/UnpinDescr object correctly.
Since the CatalogDesr has a statically allocated instance of them,
we need to make sure to copy over all settings. I didn't implement
a specific copy constructor for this, but do the copy manually depending
on the information encoded into the PinDescr/UnpinDescr instance. Here is
some room for making this more elegant.