Change UNPIN CURRENT to UNPIN PINNED.
After some discussion (jsc, avo i'm looking at you), we come to
the conclusion that CURRENT is easy to misunderstand on what
actually really gets unpinned. So change the keyword to PINNED,
since this should reasonably describe the behavior of the
current implementation (and leaves room for future improvements).

This commit also contains the tab completion code for the PIN and UNPIN
syntax.

Documentation changes are included, too.