Use sequence parser instead of expection parser.
The parser was inconsistent in its usage of expection
vs. sequence parser operators (> vs. >>).

This wasn't a problem with boost::spirit below version 1.64,
but it seems with >= 1.64 it indeed is. Since we haven't switched
to a better token representation yet, use a sequence operator consistently
throughout the whole grammar, unless we tackle the problem and
can debug it further.