First steps on parser support for CREATE RETENTION POLICY.
This commits adds first steps of the parser to learn
about the CREATE RETENTION POLICY command family.

Currently this contains rules for the new syntax

CREATE RETENTION POLICY <ident>
   { KEEP | DROP } WITH LABEL <regular expression>

This will also the first retention policies being
implemented in the catalog, as a teaching experiment and
to prove the upcoming API.

Also update the command reference according to it. This also
envisions a polished syntax for the other possible retention
rules, but this can change again if we decide to support something
more attractive.