Fix check for shmatt_t.
I was misguided to use symbol checking for shmatt_t within
CMakeLists.txt, since it is a datatype, *not* a symbol.

Use CheckTypeSize instead for cmake which returns an empty
value in case the the datatype isn't defined.

We might improve this check for other platforms, but since this
is especially for FreeBSD now, leave it at is (and probably improve it later).