Force physical replication connections.
Previously, specifying replication=database worked with PG10
to establish a physical replication connection, but starting
with PG11 this seems to have changed. It's document that you should
use replication=on|true|1 to establish a physical replication
connection now.

I haven't seen this change anywhere in the release notes and a quick
tests shows that PG10 still works with replication=database, so i'm
not entirely sure what has changed in detail.

But go with replication=true for now, as this is the documented way
in PG11.