Improve command line support for background job control.
This introduces new command line parameters --launcher and
--wal-streamer to start the background launcher and archive specific
wal streamer processes via cmdline directly.

Also add first steps for systemd unit files and integrate this
new stuff into the RHEL7 example SPEC file.

This is highly untested and subject to be improved, but commit it
for now to make it available for testing ;)