Introduce FORCE_SYSTEMID_UPDATE option for START BASEBACKUP.
When streaming a basebackup into an existing backup archive,
which already contains basebackups streamed from a PGDATA
with a different SYSTEMID, we refuse to create this basebackup
since the cluster directory is incompatible. The FORCE_SYSTEMID_UPDATE
option can be specified to override this restriction to allow
backups in such cases (e.g. if an instance was upgraded to a new
major version).