Fix wrong usage of XLogSegNoOffsetToRecPtr().
We accidently give the arguments in the wrong order (recptr has to
be the *last* argument here), which led to strange SIGFPE errors.

While digging into this bug, rename the identifyDeletionOffset()
method in ArchiveLogDirectory into checkCleanupDescriptor(), since
it doesn't do anything with the XLogRecPtr deletion offset anymore.
We will keep it as an centralized routine to check a cleanup
descriptor and its properties wether it is sane.