LabelRetention: Handle aborted and in-progress basebackup correctly.
LabelRetention::apply() was too careless when decoding
the XLOG positions read from a basebackup descriptor. This fails
with a parse exception, since the ending WAL position in such
descriptors are usually empty.

The (currently) correct behavior is to special case the WAL
positions to be the one from the starting position stored
in those descriptors. This makes sense, since those basebackups
aren't usable anyways (in case they are aborted). The "in-progress"
case is handled the same way, but those descriptors get a special
handling during the apply phase anyways and won't arrive in the
cleanup descriptors later.