Cleanup basebackup verification.
The code as-is doesn't justify an virtual method implementation
in StreamingBaseBackupDirectory. Instead, rename verify_basebackup()
into verify() and make it a static class method.

Instrument the new static method in the LIST BASEBACKUPS command, since
we need there to recognize wether a basebackup has a valid on-disk
representation or not (otherwise the code there will fail when e.g.
we want to derive its on-disk size).