Fix the way we identify XLogRecPtr cleanup offsets.
The former algorithm errornously took the xlogposend
of a basebackup as the starting point for cleanup up XLOG
segment files, which was in fact the wrong direction.

Change it so that we go backward in the XLOG archive by
starting at the XLOG start position of a basebackup. We
need to jump to the XLOG segment *before* the basebackup was taken,
which is the the entry point into the deletion process.