Fix bug in BackupCatalog::getBaseBackup().
The overloaded version with BaseBackupRetrieveMode missed
(for some really stupid reasons) to specify the archive_id,
which is required to identify a basebackup according to the
requirements encoded by the specified retrieval mode for a
basebackup.

Fix this by changing to signature of this version
for getBaseBackup() by requiring the archive_id in the signature.