First attempt to write timeline history files.
Add API methods to ArchiveLogDirectory to flush
timeline history information persistent to disk.

Add code to StartStreamingForArchiveCommand to
have the stream timeline history information written
to disc.

The code as is a first iteration and there seems
to be a lurking bug somewhere, since materialization
of the timeline history file doesn't happen.

Further commits will fix this issue.