Fix various bugs in timeline history file allocation.
During streaming, various bugs prohibited a correct allocation
of the timeline history file in the archive.

The code as-is is not safe yet, e.g. we need to test against
upstream servers which had multiple timeline switches during
outages of the walstreamer. This should show wether the current
code in the WALStreamer loop can safely keep up.

Follow up commits to this issue likely ahead ;)