Rely on GNU implementation for strerror().
Commit 26e9d4d4ef1 in PostgreSQL upstream exposes the
pg_strerror() and pg_strerror_r() API to clients, which is not
workable when using boost (or at least boost::interprocess).

The definitions in /usr/include/boost/interprocess/errors.hpp
(specifically boost::interprocess::fill_system_message) collides
with the ifdef'ed redefinition of strerror() from PostgreSQL's
port.h, which is included by postgres_fe.h we need for some type
definitions. PostgreSQL upstream wants frontend clients to rely
on their exposed error API, but since we use boost with it's own
infrastructure around we need to undef the imported definition from
PostgreSQL to get back to a at least compilable and workable
environment.

This might not be the proper solution yet, but for now it seems
to me the right choice.