Fix CountRetention::drop_num() to not drop every valid basebackup.
In case a DROP <num> retention rule was specified, where <num> will
be higher than the current number of valid basebackups in the archive and
this retention policy is applied to this specific archive, error out.

We don't want to have the user accidently dropping all valid basebackups from
an archive, and also don't want to be silent if such a thing happens. This
makes the DROP +<num> retention rules somehow different to others, since
in this case we will throw. But from a user perspective this seems more sane,
preventing them from dropping all available backups. This way at least the recent
one stays available and the user gets an error, telling him he'd dropped
all his life savers...