Introduce new static method CPGBackupCtlBase::resolve_file_path().
This static method checks wether the given executable name as string
can be found in PATH. Currently this is only portable between Unices, but
since we don't support Windows atm this isn't problemtic for now.

Employ this new method when creating a backup profile with either
COMPRESSION PLAIN or COMPRESSION PBZIP, to make sure the binary can
be found and the backup process successfully started.