Fix uncautious usage of StreamingBackupDirectory::size().
When listing basebackups via LIST BASEBACKUP IN ARCHIVE in
compact mode, we were a little careless when getting the
size of the streaming backup directory. If the basebackup status
is not BASEBACKUP_OK, this will throw.

Move the display code into proper if clauses to make sure we
don't stumble accross invalid streaming backup directories.