Introduce START RECOVERY STREAM and infrastructure.
This commit implements a starter for boost::asio supported
streaming instances, which allows to stream archived basebackups
over the wire. This also means to move headers around. The logic
for server and clients is located within its own new shared
library libpgbckctl-proto.so.

It doesn't work right now, merely a commit to save the current
work.