Fix upstream reporting of WALStreamer flush position.
Because the coding was a little too lax, we missed to report
the correct flush location at the end of a XLOG segment file.
This lead to a bloated pg_wal directory on upstream instances,
caused by never updating the restart_lsn of the used replication
slot by a WALStreamer.

Change the way we set the current flush position by adding an
additional parameter to TransactionLogBackup::write() archive handler,
which is responsible to close, create and flush current XLOG segment
files in the archive. The parameter will be set to a valid XLogRecPtr
value, indicating the position the current WAL stream was flushed. In
detail, this is normally the starting offset for the new XLOG segment.

Also force WALStreamer::sendStatusUpdate() to always send the flush
position by calling ReceiverStatusMessage::reportFlushPosition().