Further work on PG protocol implementation.
Add an ProtocolErrorStack() implementation to ease sending
ErrorResponse messages over the wire.

Add a stubb for TarRecovery (mostly for brainstorming purposes):

Also teach the PGProtoStreamingServer about our new error stack,
but this work is not yet complete.

This commit also includes a fix to change the datatype of tablespace
OIDs from unsigned integer to signed integer. This should fix a long
standing bug when writing the OID as filename to the archive (see also
issue #11).