Move logging from poor man's std::cout/std:cerr to boost::log::trivial.
Use boost::log::trivial where appropiate and drop most of the current usage
of std::cerr and std::cout. For now we keep the __DEBUG__ and __DEBUG_XLOG__ macros
which special case the occurence for DEBUG messages, but i plan to incorporate that
mess also into the RuntimeEnvironment class.

There's still the question on how to do that, so leave that for further work.