Commit various work on PostgreSQL protocol support.
Add protocol result sets to create and return result sets to the
client (simple protocol).

Implement a parser for PostgreSQL streaming protocol commands with our
own extensions (here: LIST_BASEBACKUPS) and the infrastructure around this.
This includes a command handler infrastructure similar to the PGBackupCtlParser
and friends.

Nothing really fully workable and full of bugs, but commit it now nevertheless
to not loose any changes.