Add centralized backup lock info infrastructure.
Central object is a BackupLockInfoAggregator class, aggregating a BackupLockInfo
descendant class which encapsulated lock information of a specific lock type.

Currently present are BackupPinnedValidLockInfo, which tells the caller wether
a basebackup was either pinned or is invalid (this is a scary notion for "locked"
basebackups, but we consider in-progress or aborted basebackups as locked, for example
to prevent retention policies to delete WAL files unconditionally in such cases).

The other one is the SHMBackupLockInfo which provides information if a specific
basebackup was locked by a background worker child process (via child_info).

The interface for BackupLockInfo classes is fairly simple, there is an abstract
BackupLockInfo::locked() method which must be implemented by each descendant.

Employ the BackLockInfo infrastructure in Retention classes, which will replace the old
manual lock checking by having a BackupLockInfoAggregator instead.

There's some tricky hack involved, since the APPLY RETENTION POLICY command
is not allowed to check for SHMBackupLockInfo in case _no_ launcher was established.
In this case no background shared memory segment is created. So check for any launcher
running, which will be done by the new credativ::launcher_is_running() function. We
might have this new function also do its duty in the SHOW WORKER command handler, but
i leave this for further improvements.