Implement PostgreSQL streaming command TIMELINE_HISTORY.
This allows to return a timeline history file according to
the PostgreSQL streaming replication protocol.

This commit also changes the way to access catalog and filesystem
for streaming replication commands. Usually this is done via a
PGProtoCatalogHandler instance, which was employed by each protocol
streaming command. This meant to setup catalog each time we want to
execute a command, which annoyed me during implementing the TIMELINE_HISTORY
command.

Thus, move the initialization to the backend startup procedure and
initialize a PGProtoCatalogHandler only once, when we setup our
recovery instance during connection startup. This makes the whole
thing easier to maintain and a lot more flexible in my opinion.

Also fix BackupFile descendants to correctly handle feof() and ferror()
conditions (looks like i forgot this a while ago, ArchivePipedFile already
does it right).