Teach StreamBaseBackup objects to operate READ/WRITE.
In preparation for the restore/recovery infrastructure, we want
to instruct StreamBaseBackup to exactly know what the difference between
backup and recovery is. I don't want to have a specific class for this,
my idea is to use the StreamBaseBackup as a multiplexing handler between
the backup and restore processes. The new SB_WRITE and SB_READ status will
reflect those operation modes (SB_WRITE for backup, SB_READ for restore).