Enable Boost.Test framework.
This allows us to write unit tests against pg_backup_ctl++ and its
libraries. We also need to have a testing SQLite database during
the tests, so add a custom command and custom target dependency to
libpgbckctl-common.

Unit tests aren't enabled by default, you have to specify

-DBUILD_UNIT_TESTS=1

to cmake to enable unit tests. This is not the final solution though, because
it's ugly and cmake complains about the otherwise unused
variable BUILD_UNIT_TESTS. Instead, this will be changed to a cmake feature
definition later.