Introduce special connection database identifiers for recovery instance.
"latest" and "newest" can be used in the connection DSN as the
database name to attach to the latest basebackup in the archive (if any).

"oldest" references the oldest basebackup.

Also remove the exception thrown in BackupCatalog::close() in case the
catalog handle was already closed. This confuses callers more than it does
any good. We treat multiple closes on a catalog handle as a no-op
from now on.