Change the way process_startup_gucs() handles catalog errors.
Previously we had the idea to just return specific error codes in
case something goes wrong when attaching a requested basebackup
identifier to a streaming server context. This turned out to be error
prone in case you miss to correctly handle thrown exceptions.

So better rely on the exceptions thrown by catalog access and do the
right thing by just catching them. Currently this is a CCatalogIssue
exception only, but this might change in the future.

This way we just need to catch the exceptions in process_startup_gucs()
only, send an error message to the client and we're done.