Fix handling of backup manifest during basebackup stream.
The contents of a backup manifest are streamed after having
received the tar archive, containing the files of the backup itself.
This content is received in plain format, so make sure we handle it
accordingly. This involves switching the compression mode of the backup
handler (an instance of StreamBaseBackup) to BACKUP_COMPRESS_TYPE_NONE
temporarily, so that it allocates a plain ArchiveFile without compression
or piping.