Introduce the concept of executable contexts for streaming commands.
This implements an ExecutableContext base class which can be used to
create command-specific context classes for their execution. This for
example allows to specify specific runtime environments for
PGProtoStreamingCommands and their execute() method. An ExecutableContext
instance is passed as a shared pointer to execute(), carrying various
information specified by the executing recovery instance.

In the long term this should make specifying runtime information in the
recovery instance' RuntimeConfiguration obsolete, but this commit just
introduces the ExecutableContext API for now.