Suppress warnings for misuse of bind placeholders in global namespace.
Newer boost version (>= 1.68 i think) want to have bind placeholders
within boost::placeholder namespace, not in the global one. As long as
i'm too lazy to fix this, suppress the warning by specifying
BOOST_BIND_GLOBAL_PLACEHOLDERS during compilation.

Before this is going to be removed, we certainly need to fix this.