Fix wrong handling of output runtime variable list_archive.mode.
Sadly, we need to explicitely specify the target type we want
to have for ConfigValue string. Otherwise the target value gets
routed to its boolean sister, causing it to be casted to "true" boolean
values, which isn't what we want.

I've somehow missed this issue when designing the RuntimeConfiguration
facility. To circumvent this gotcha, always instantiate a std::string
object explicitely.