Fix aftermath of merge commit abb5b728bc31c1437182b0938f2c7bc80f505550.
This commit refactors BackupDirectory::verify() and friends into
a bool returning function exists(). However, while adding exists(path)
as a protected method to BackupDirectory, we also changed the scope of
boost::filesystem::exists(), formerly used by StreamingBaseBackupDirectory::create()
to our very own exists() call, which is not the right condition
to comply.

It throws if the specified path handle does not exist,
which is the exact opposite outcome we want here.
START BASEBACKUP will crash in this case, since it doesn't create
the streaming subdirectory as expected.

So, change the scope to explicit call boost::filesystem::exists(), like in the
other places. We don't need that though, since this commit also
renames BackupDirectory::exists(path) to checkPathForArchive(),
since that's the right name reflecting what it's real purpose
is. But it looks more descriptive in this case, since during debugging
the issue above i've had a hard time to parse which exists() was
meant in reality.

Someone might be tempted to say "well name it differently",
but exists() is still the thing we want (especially in BackupDirectory ancestors
like ArchiveLogDirectory), so keep the structure, at least for now.