Overhaul error handling in pg_backup_ctl++ command parser.
Formerly, we didn't employ expectation points (EPS) in our boost:spirit
based parser. With this commit, we introduce them to improve the error
handling by allowing to put a meaningful parser message in a CParserIssue
exception, thrown when encountering a parser error.

Formerly, just an error condition was set telling the caller that
something went wrong. But we didn't say *what* exactly went wrong when
parsing the command. With EPS we have a better message handling in our
hands. For example, the following command will not be parsed since the
keyword "BASEBACKUP" is misspelled (the plural is missing):

   LIST BASEBACKUP IN ARCHIVE pg13 ;

Formerly we would have gotten just a message, that the command couldn't
be parsed. Now the parser tells exactly what he expects instead:

   expecting <alternative><ARCHIVE><BACKUP CATALOG><CONNECTION><BASEBACKUPS><RETENTION> here: "BASEBACKUP IN ARCHIVE pg13"

This is currently not perfect, but much better what we had before.

Also fix various parser bugs and more.